server/python/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
rembg>=2.0.50
pillow>=10.0.0
openai-whisper>=20231117
onnxruntime>=1.15.0
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
super-image>=0.1.7
